Organize write module
